requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
python-dotenv
fpdf